// Capstone Project/src/app/pages/authentication/signin-page/screen/types.ts
export enum authenticationEnum {
  SIGN_IN_REQUEST = 'APP/AUTHENTICATION/SIGN_IN_REQUEST',
  SIGN_IN_SUCCESS = 'APP/AUTHENTICATION/SIGN_IN_SUCCESS',
  SIGN_IN_FAILURE = 'APP/AUTHENTICATION/SIGN_IN_FAILURE',

  SIGN_IN_WITH_GG_REQUEST = 'APP/AUTHENTICATION/SIGN_IN_WITH_GG_REQUEST',
  SIGN_IN_WITH_GG_SUCCESS = 'APP/AUTHENTICATION/SIGN_IN_WITH_GG_SUCCESS',
  SIGN_IN_WITH_GG_FAILURE = 'APP/AUTHENTICATION/SIGN_IN_WITH_GG_FAILURE',

  SIGN_IN_GET_USER_INFOR_REQUEST = 'APP/AUTHENTICATION/SIGN_IN_GET_USER_INFOR_REQUEST',
  SIGN_IN_GET_USER_INFOR_SUCCESS = 'APP/AUTHENTICATION/SIGN_IN_GET_USER_INFOR_SUCCESS',
  SIGN_IN_GET_USER_INFOR_FAILURE = 'APP/AUTHENTICATION/SIGN_IN_GET_USER_INFOR_FAILURE',

  FORGOT_PASSWORD_REQUEST = 'APP/AUTHENTICATION/FORGOT_PASSWORD_REQUEST',
  FORGOT_PASSWORD_SUCCESS = 'APP/AUTHENTICATION/FORGOT_PASSWORD_SUCCESS',
  FORGOT_PASSWORD_FAILURE = 'APP/AUTHENTICATION/FORGOT_PASSWORD_FAILURE',

  RESEND_EMAIL_VERIFY_REQUEST = 'APP/AUTHENTICATION/RESEND_EMAIL_VERIFY_REQUEST',
  RESEND_EMAIL_VERIFY_SUCCESS = 'APP/AUTHENTICATION/RESEND_EMAIL_VERIFY_SUCCESS',
  RESEND_EMAIL_VERIFY_FAILURE = 'APP/AUTHENTICATION/RESEND_EMAIL_VERIFY_FAILURE',

  CLEAR_SIGN_IN_STATE = 'APP/AUTHENTICATION/SIGN_IN_CLEAR_STATE',
}